requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-vision==2.3.1
google-cloud-storage==1.38.0
google-cloud-firestore==2.1.1